Test organization clean up
- simplify pytest imports
- move logged_in_admin_client to root conftest
