+++ bikespace_api/bikespace_api/admin/tests/test_admin_page.py
@@ -2,23 +2,12 @@ import re
 
 import pytest
 from bs4 import BeautifulSoup
-from pytest import mark
 
 from bikespace_api.admin.roles import ApplicationRoles
 
 # Rationale for 'type: ignore' comments:
 # - soup.find doesn't have an overload where both name= and string= are not None (as of bs4 version 4.14.2), even though this is a documented usage of the function
 # - soup.find string= parameter is not typed for regular expressions even though these are a documented input type
-
-
-@pytest.fixture()
-def logged_in_admin_client(test_client, clean_db):
-    test_client.post(
-        "/admin/login/",
-        data=dict(email="admin@example.com", password="admin"),
-        follow_redirects=True,
-    )
-    return test_client
 
 
 def test_admin_page(test_client):
@@ -115,7 +104,7 @@ def default_login_page_redirect(response, soup: BeautifulSoup):
     )
 
 
-@mark.uses_db
+@pytest.mark.uses_db
 def test_admin_login_successfully(test_client, clean_db):
     """
     GIVEN the flask application configured for testing
@@ -150,7 +139,7 @@ def test_admin_login_successfully(test_client, clean_db):
     assert post_login_soup.find(string=re.compile("Admin"))
 
 
-@mark.uses_db
+@pytest.mark.uses_db
 def test_create_and_update_a_user(logged_in_admin_client):
     """
     GIVEN the flask application configured for testing
@@ -208,7 +197,7 @@ def test_create_and_update_a_user(logged_in_admin_client):
     )
 
 
-@mark.uses_db
+@pytest.mark.uses_db
 def test_admin_submission_create_form_renders(logged_in_admin_client):
     """
     GIVEN a logged-in superuser
@@ -221,7 +210,7 @@ def test_admin_submission_create_form_renders(logged_in_admin_client):
     assert soup.find("form")
 
 
-@mark.uses_db
+@pytest.mark.uses_db
 def test_admin_roles_accessible_as_superuser(logged_in_admin_client):
     """
     GIVEN a logged-in superuser
@@ -234,7 +223,7 @@ def test_admin_roles_accessible_as_superuser(logged_in_admin_client):
     assert soup.find("table")
 
 
-@mark.uses_db
+@pytest.mark.uses_db
 def test_update_user_without_changing_password(logged_in_admin_client):
     """
     GIVEN a logged-in superuser and an existing user
@@ -303,7 +292,7 @@ def test_update_user_without_changing_password(logged_in_admin_client):
         assert verify_password(original_password, user.password)
 
 
-@mark.uses_db
+@pytest.mark.uses_db
 def test_allowed_pages_for_regular_users(test_client, clean_db):
     """
     GIVEN the flask application configured for testing and
